+++ mDeepFRI/alignment.py
@@ -1,103 +1,125 @@
 import logging
+from collections import defaultdict
 from functools import partial
-from multiprocessing.pool import ThreadPool
+from multiprocessing.pool import Pool
 
 import numpy as np
 import pyopal
 
-from mDeepFRI.bio_utils import (AlignmentResult, load_query_sequences,
+from mDeepFRI.bio_utils import (AlignmentResult, load_fasta_as_dict,
                                 retrieve_fasta_entries_as_dict)
 from mDeepFRI.mmseqs import (filter_mmseqs_results, run_mmseqs_search,
                              validate_mmseqs_database)
 
 logger = logging.getLogger(__name__)
 
-
-def align_best_score(query_item: str, database: pyopal.Database,
-                     target_names: list, gap_open: int, gap_extend: int,
-                     identity_threshold: float):
-
-    query_name, query_sequence = query_item
-
-    logger.debug("Aligning %s", query_name)
-    # current bug in pyOpal
-    # https://github.com/althonos/pyopal/issues/3
-    try:
-        results = database.search(query_sequence,
-                                  mode="score",
-                                  algorithm="nw",
-                                  gap_open=gap_open,
-                                  gap_extend=gap_extend)
-
-        best = max(results, key=lambda x: x.score)
-        best_sequence = database[best.target_index]
-        best_name = target_names[best.target_index]
-        single_seq = pyopal.Database([best_sequence])
-        alignment = single_seq.search(query_sequence,
-                                      mode="full",
-                                      algorithm="nw",
-                                      gap_open=gap_open,
-                                      gap_extend=gap_extend)[0].alignment
-
-        result = AlignmentResult(query_name, query_sequence, best_name,
-                                 best_sequence, alignment)
-        result.insert_gaps().calculate_identity()
-        logger.debug("Alignment of %s finished.", query_name)
-        if result.identity >= identity_threshold:
-            pass
-        else:
-            logger.debug("No alignment for  %s.", query_name)
-            result = None
-    except RuntimeError:
-        logger.debug("Alignment of %s failed.", query_name)
-        result = None
-
-    return result
+##TODO: do not collect large database
+##TODO: align 1 query against top-k databases
 
 
-def align_query(query_seqs: dict,
-                target_seqs: dict,
-                alignment_gap_open: float,
-                alignment_gap_extend: float,
-                identity_threshold: str = 0.3,
-                threads: int = 1):
+def best_hit_database(query,
+                      target_sequences,
+                      gap_open: int = 10,
+                      gap_extend: int = 1):
+    """
+    Find the best hit in the database and return index.
 
-    logger.info("Pairwise alignment started.")
-    logger.info("Aligning %i queries against %i database sequences.",
-                len(query_seqs), len(target_seqs))
-    # Default substitution matrix BLOSUM50
-    database = pyopal.Database(list(target_seqs.values()))
-    align_against_db = partial(align_best_score,
-                               database=database,
-                               target_names=list(target_seqs.keys()),
-                               gap_open=alignment_gap_open,
-                               gap_extend=alignment_gap_extend,
-                               identity_threshold=identity_threshold)
+    Args:
+        query (str): The query sequence.
+        target_sequences (dict): The target sequences.
 
-    with ThreadPool(threads) as pool:
-        all_alignments = pool.map(align_against_db, query_seqs.items())
+    Returns:
+        int: The index of the best hit in the database.
+    """
 
-    filtered_alignments = list(filter(None, all_alignments))
-    logger.info("Found %i alignments.", len(filtered_alignments))
-    logger.info("Pairwise alignment finished.")
+    aligner = pyopal.Aligner(gap_open=gap_open, gap_extend=gap_extend)
+    target_database = pyopal.Database(target_sequences.values())
 
-    return filtered_alignments
+    # Retrieve the best hit
+    results = aligner.align(query,
+                            target_database,
+                            mode="score",
+                            overflow="buckets",
+                            algorithm="nw")
+    best_hit = max(results, key=lambda x: x.score)
+    best_index = best_hit.target_index
+    best_idx = list(target_sequences.keys())[best_index]
+    best_seq = target_sequences[best_idx]
+
+    return best_idx, best_seq
+
+
+def align_pairwise(query, target, gap_open: int = 10, gap_extend: int = 1):
+    """
+    Aligns the query against the target and returns the alignment.
+
+    Args:
+        query (str): The query sequence.
+        target (str): The target sequence.
+        aligner (pyopal.Aligner): The aligner object.
+
+    Returns:
+        pyopal.Alignment: The alignment of the query against the target.
+    """
+
+    aligner = pyopal.Aligner(gap_open=gap_open, gap_extend=gap_extend)
+    database = pyopal.Database([target])
+    # Align the sequences
+    alignment = aligner.align(query, database, algorithm="nw", mode="full")
+
+    return alignment[0].alignment
+
+
+def pairwise_against_database(query_id,
+                              query_sequence,
+                              target_sequences,
+                              gap_open: int = 10,
+                              gap_extend: int = 1):
+    """
+    Finds the best alignment of the query against the target.
+    """
+
+    # find best hits in the database
+    best_idx, best_target = best_hit_database(query_sequence, target_sequences,
+                                              gap_open, gap_extend)
+    # align the query against the best hit
+    alignment = align_pairwise(query_sequence, best_target, gap_open,
+                               gap_extend)
+    # create an alignment object
+    alignment_result = AlignmentResult(query_id, query_sequence, best_idx,
+                                       best_target, alignment)
+
+    return alignment_result
+
+
+def create_partial_database(top_matches, target_sequences):
+    """
+    Create a partial database from the target sequences.
+
+    Args:
+        query_id (str): The
+    """
+
+    partial_db = {k: target_sequences[k] for k in top_matches}
+
+    return partial_db
 
 
 def run_alignment(query_file: str,
                   mmseqs_db: str,
                   sequence_db: str,
                   output_path: str,
-                  mmseqs_min_bit_score: int = None,
+                  mmseqs_min_bitscore: int = None,
                   mmseqs_max_eval: float = 10e-5,
                   mmseqs_min_identity: float = 0.3,
-                  top_k: int = 30,
+                  top_k: int = 5,
                   alignment_gap_open: int = 10,
-                  alignment_gap_continuation: int = 1,
-                  identity_threshold: float = 0.5,
+                  alignment_gap_extend: int = 1,
                   threads: int = 1):
 
-    query_seqs = load_query_sequences(query_file, output_path)
+    align_partial = partial(pairwise_against_database,
+                            gap_open=alignment_gap_open,
+                            gap_extend=alignment_gap_extend)
     mmseqs_valid = validate_mmseqs_database(mmseqs_db)
     if not mmseqs_valid:
         raise ValueError(f"MMSeqs2 database not found in {mmseqs_db}.")
@@ -105,11 +127,17 @@ def run_alignment(query_file: str,
     mmseqs_results = run_mmseqs_search(query_file, mmseqs_db, output_path,
                                        threads)
     filtered_mmseqs_results = filter_mmseqs_results(mmseqs_results,
-                                                    mmseqs_min_bit_score,
+                                                    mmseqs_min_bitscore,
                                                     mmseqs_max_eval,
                                                     mmseqs_min_identity,
                                                     k_best_hits=top_k,
                                                     threads=threads)
+
+    query_seqs = load_fasta_as_dict(query_file)
+
+    unique_queries = defaultdict(list)
+    for result in filtered_mmseqs_results:
+        unique_queries[result[0]].append(result[1])
 
     # if MMSeqs2 alignment is empty
     if filtered_mmseqs_results is None:
@@ -117,9 +145,17 @@ def run_alignment(query_file: str,
     else:
         target_ids = np.unique(filtered_mmseqs_results["target"]).tolist()
         target_seqs = retrieve_fasta_entries_as_dict(sequence_db, target_ids)
+        # create partial databases
+        partial_databases = {
+            k: create_partial_database(v, target_seqs)
+            for k, v in unique_queries.items()
+        }
 
-        alignments = align_query(query_seqs, target_seqs, alignment_gap_open,
-                                 alignment_gap_continuation,
-                                 identity_threshold, threads)
+        # align the query against the best hit
+        with Pool(threads) as pool:
+            alignments = pool.starmap(
+                align_partial,
+                [(query_id, query_seqs[query_id], partial_databases[query_id])
+                 for query_id in unique_queries])
 
     return alignments