+++ mDeepFRI/bio_utils.py
@@ -1,62 +1,79 @@
-import json
-import logging
 from io import StringIO
-from pathlib import Path
 from typing import Dict, List, Tuple
 
 import foldcomp
 import numpy as np
 import requests
-from Bio.PDB import MMCIFParser, PDBParser
+from biotite.sequence import ProteinSequence
+from biotite.structure.io.pdb import PDBFile
+from biotite.structure.io.pdbx import PDBxFile, get_structure
 from pysam import FastaFile, FastxFile
-from scipy.spatial.distance import pdist, squareform
 
-from mDeepFRI.alignment_utils import alignment_identity
+from mDeepFRI.alignment_utils import alignment_identity, pairwise_sqeuclidean
 
-# copied from Biopython to remove dependency
-protein_letters_1to3 = {
-    "A": "Ala",
-    "C": "Cys",
-    "D": "Asp",
-    "E": "Glu",
-    "F": "Phe",
-    "G": "Gly",
-    "H": "His",
-    "I": "Ile",
-    "K": "Lys",
-    "L": "Leu",
-    "M": "Met",
-    "N": "Asn",
-    "P": "Pro",
-    "Q": "Gln",
-    "R": "Arg",
-    "S": "Ser",
-    "T": "Thr",
-    "V": "Val",
-    "W": "Trp",
-    "Y": "Tyr",
-}
-protein_letters_1to3_extended = {
-    **protein_letters_1to3,
-    **{
-        "B": "Asx",
-        "X": "Xaa",
-        "Z": "Glx",
-        "J": "Xle",
-        "U": "Sec",
-        "O": "Pyl"
-    },
-}
 
-protein_letters_3to1_extended = {
-    value: key
-    for key, value in protein_letters_1to3_extended.items()
-}
+class AlignmentResult:
+    """
+    Class for storing pairwise alignment results.
 
-PROTEIN_LETTERS = dict()
-for k, v in protein_letters_3to1_extended.items():
-    PROTEIN_LETTERS[str.upper(k)] = v
-PROTEIN_LETTERS["UNK"] = "X"
+    Attributes:
+        query_name (str): Name of the query sequence.
+        query_sequence (str): Query sequence.
+        target_name (str): Name of the target sequence.
+        target_sequence (str): Target sequence.
+        alignment (str): Alignment string.
+        gapped_sequence (str): Query sequence with gaps.
+        gapped_target (str): Target sequence with gaps.
+        identity (float): Identity between two sequences.
+
+    Methods:
+        insert_gaps: Inserts gaps into query and target sequences.
+        calculate_identity: Calculates identity between query and target.
+    """
+    def __init__(self, query_name, query_sequence, target_name,
+                 target_sequence, alignment):
+
+        self.query_name = query_name
+        self.query_sequence = query_sequence
+        self.target_name = target_name
+        self.target_sequence = target_sequence
+        self.alignment = alignment
+        self.insert_gaps()
+        self.calculate_identity()
+        self.db_name = None
+
+    def __str__(self):
+        return f"AlignmentResult(query_name={self.query_name}, target_name={self.target_name}, " \
+               f"identity={self.identity})"
+
+    def __repr__(self):
+        return f"AlignmentResult(query_name={self.query_name}, target_name={self.target_name}, " \
+               f"identity={self.identity})"
+
+    def insert_gaps(self):
+        """
+        Inserts gaps into query and target sequences.
+        """
+
+        self.gapped_sequence, self.gapped_target = insert_gaps(
+            self.query_sequence, self.target_sequence, self.alignment)
+
+        return self
+
+    def calculate_identity(self):
+        """
+        Calculates identity between query and target.
+
+        Args:
+            None
+        Returns:
+            self
+        """
+
+        self.identity = alignment_identity(self.gapped_sequence,
+                                           self.gapped_target)
+
+        return self
 
 
 def seq2onehot(seq: str) -> np.ndarray:
@@ -114,91 +131,6 @@ def insert_gaps(sequence: str, reference: str,
     return "".join(sequence), "".join(reference)
 
 
-class AlignmentResult:
-    """
-    Class for storing pairwise alignment results.
-
-    Attributes:
-        query_name (str): Name of the query sequence.
-        query_sequence (str): Query sequence.
-        target_name (str): Name of the target sequence.
-        target_sequence (str): Target sequence.
-        alignment (str): Alignment string.
-        gapped_sequence (str): Query sequence with gaps.
-        gapped_target (str): Target sequence with gaps.
-        identity (float): Identity between two sequences.
-
-    Methods:
-        insert_gaps: Inserts gaps into query and target sequences.
-        calculate_identity: Calculates identity between query and target.
-    """
-    def __init__(self,
-                 query_name,
-                 query_sequence,
-                 target_name,
-                 target_sequence,
-                 alignment,
-                 gapped_sequence="",
-                 gapped_target="",
-                 identity=None):
-
-        self.query_name = query_name
-        self.query_sequence = query_sequence
-        self.target_name = target_name
-        self.target_sequence = target_sequence
-        self.alignment = alignment
-        self.gapped_sequence = gapped_sequence
-        self.gapped_target = gapped_target
-        self.identity = identity
-        self.db_name = None
-
-    def __str__(self):
-        return f"AlignmentResult(query_name={self.query_name}, target_name={self.target_name}, " \
-               f"identity={self.identity})"
-
-    def __repr__(self):
-        return f"AlignmentResult(query_name={self.query_name}, target_name={self.target_name}, " \
-               f"identity={self.identity})"
-
-    def insert_gaps(self):
-        """
-        Inserts gaps into query and target sequences.
-
-        Args:
-            None
-        Returns:
-            self
-        """
-        seq = list(self.query_sequence)
-        ref = list(self.target_sequence)
-        aln = list(self.alignment)
-
-        for i, a in enumerate(aln):
-            if a == "I":
-                seq.insert(i, "-")
-            elif a == "D":
-                ref.insert(i, "-")
-        self.gapped_sequence = "".join(seq)
-        self.gapped_target = "".join(ref)
-
-        return self
-
-    def calculate_identity(self):
-        """
-        Calculates identity between query and target.
-
-        Args:
-            None
-        Returns:
-            self
-        """
-
-        self.identity = alignment_identity(self.gapped_sequence,
-                                           self.gapped_target)
-
-        return self
-
-
 def load_fasta_as_dict(fasta_file: str) -> Dict[str, str]:
     """
     Load FASTA file as dict of headers to sequences
@@ -207,11 +139,12 @@ def load_fasta_as_dict(fasta_file: str) -> Dict[str, str]:
         fasta_file (str): Path to FASTA file. Can be compressed.
 
     Returns:
-        Dict[str, str]: Dictionary of FASTA entries.
+        Dict[str, str]: Dictionary of FASTA entries sorted by length.
     """
 
     with FastxFile(fasta_file) as fasta:
         fasta_dict = {entry.name: entry.sequence for entry in fasta}
+
     return fasta_dict
 
 
@@ -235,65 +168,6 @@ def retrieve_fasta_entries_as_dict(fasta_file: str,
     return fasta_dict
 
 
-def load_query_sequences(query_file: str, output_path: str) -> Dict[str, str]:
-    """
-    Loads query protein sequences from FASTA file. Filters
-    out sequences that are too short or too long.
-
-    Args:
-        query_file (str): Path to FASTA file with query protein sequences.
-        output_path (str): Path to output folder.
-
-    Returns:
-        query_seqs (dict): Dictionary with query protein headers to sequences.
-    """
-
-    # By DeepFRI design (60, 1000)
-    MIN_PROTEIN_LENGTH = 60
-    MAX_PROTEIN_LENGTH = 1_000
-
-    query_file = Path(query_file)
-    output_path = Path(output_path)
-
-    query_seqs = load_fasta_as_dict(query_file)
-
-    if len(query_seqs) == 0:
-        raise ValueError(
-            f"{query_file} does not contain parsable protein sequences.")
-
-    logging.info("Found total of %i protein sequences in %s", len(query_seqs),
-                 query_file)
-
-    # filter out sequences that are too short or too long
-    prot_len_outliers = {}
-    for prot_id, sequence in query_seqs.items():
-        prot_len = len(sequence)
-        if prot_len > MAX_PROTEIN_LENGTH or prot_len < MIN_PROTEIN_LENGTH:
-            prot_len_outliers[prot_id] = prot_len
-
-    for outlier in prot_len_outliers.keys():
-        query_seqs.pop(outlier)
-
-    # write skipped protein ids to file
-    if len(prot_len_outliers) > 0:
-        logging.info(
-            "Skipping %i proteins due to sequence length outside range %i-%i aa.",
-            len(prot_len_outliers), MIN_PROTEIN_LENGTH, MAX_PROTEIN_LENGTH)
-        logging.info("Skipped protein ids will be saved in " \
-                     "metadata_skipped_ids_length.json.")
-        json.dump(prot_len_outliers,
-                  open(output_path / 'metadata_skipped_ids_due_to_length.json',
-                       "w",
-                       encoding="utf-8"),
-                  indent=4,
-                  sort_keys=True)
-
-    assert len(query_seqs
-               ) > 0, "All proteins were filtered out due to sequence length."
-
-    return query_seqs
-
-
 def calculate_contact_map(coordinates: np.ndarray,
                           threshold=6.0,
                           mode="matrix") -> np.ndarray:
@@ -309,8 +183,10 @@ def calculate_contact_map(coordinates: np.ndarray,
     Returns:
         np.ndarray: Contact map.
     """
+    # squared euclidean is used for efficiency
+    threshold = threshold**2
 
-    distances = squareform(pdist(coordinates))
+    distances = pairwise_sqeuclidean(coordinates)
     cmap = (distances < threshold).astype(np.int32)
 
     if mode == "sparse":
@@ -321,10 +197,8 @@ def calculate_contact_map(coordinates: np.ndarray,
     return cmap
 
 
-def extract_residues_coordinates(
-        structure_string: str,
-        chain: str = None,
-        max_seq_len: int = 1_000) -> Tuple[List, np.ndarray]:
+def extract_residues_coordinates(structure_string: str,
+                                 chain: str = None) -> Tuple[List, np.ndarray]:
     """
     Extracts residues and coordinates from structural file.
     Automatically processes PDB and mmCIF files.
@@ -336,44 +210,25 @@ def extract_residues_coordinates(
     Returns:
         Tuple[List, np.ndarray]: Tuple of residues and coordinates.
     """
-
     try:
-        parser = MMCIFParser()
-        structure = parser.get_structure("", StringIO(structure_string))
-        if chain:
-            structure = structure[0][chain]
+        mmcif = PDBxFile.read(StringIO(structure_string))
+        structure = get_structure(mmcif, model=1)
+    except UnboundLocalError:
+        pdb = PDBFile(StringIO(structure_string))
+        structure = pdb.get_structure()[0]
 
-        residues = list(structure.get_residues())
-        # get CA atom coordinates for mmCIF
-        coords = np.array([
-            atom.get_coord() for atom in structure.get_atoms()
-            if atom.get_name() == "CA"
-        ])
+    protein_chain = structure[structure.chain_id == chain]
+    # extract CA atoms coordinates
+    ca_atoms = protein_chain[protein_chain.atom_name == "CA"]
+    residues = str(ProteinSequence(ca_atoms.res_name))
+    coords = ca_atoms.coord
 
-    except KeyError:
-        parser = PDBParser()
-        structure = parser.get_structure("", StringIO(structure_string))
-        if chain:
-            structure = structure[0][chain]
-
-        residues = [x for x in structure.get_residues()][:max_seq_len]
-        coords = np.array([residue["CA"].get_coord() for residue in residues])
-
-    #     parser = MMCIFParser()
-    #     structure = parser.get_structure("", StringIO(structure_string))
-    #     if chain:
-    #         structure = structure[0][chain]
-
-    #     residues = list(structure.get_residues())
-    #     # get CA atom coordinates for mmCIF
-    #     coords = np.array([atom.get_coord() for atom in structure.get_atoms() if atom.get_name() == "CA"])
     return (residues, coords)
 
 
-def retrieve_structure_features(
-        idx: str,
-        database_path: str = "pdb100",
-        max_seq_len: int = 1_000) -> Tuple[List, np.ndarray]:
+def retrieve_structure_features(idx: str,
+                                database_path: str = "pdb100"
+                                ) -> Tuple[List, np.ndarray]:
     """
     Retrieves structure either from PDB or supplied FoldComp database.
     Extracts sequence and coordinate infromaton.
@@ -408,9 +263,7 @@ def retrieve_structure_features(
                 for _, pdb in db:
                     structure = pdb
 
-    residues, coords = extract_residues_coordinates(structure,
-                                                    chain=chain,
-                                                    max_seq_len=max_seq_len)
+    residues, coords = extract_residues_coordinates(structure, chain=chain)
 
     return (residues, coords)
 
@@ -549,9 +402,7 @@ def retrieve_align_contact_map(
         database = "pdb100"
 
     idx = alignment.target_name.rsplit(".", 1)[0]
-    coordinates = retrieve_structure_features(idx,
-                                              database_path=database,
-                                              max_seq_len=max_seq_len)[1]
+    coordinates = retrieve_structure_features(idx, database_path=database)[1]
 
     cmap = calculate_contact_map(coordinates,
                                  threshold=threshold,